Synthetic context schemes now consider racemic bracket assignments of reactants and products: Implemented IsRacemic fields for tblExperiments and backlog assignments when moving to version 4.1.0

+++ ElnCoreModel/ELNModel/tblAuxiliaries.cs
@@ -9,7 +9,7 @@ public partial class tblAuxiliaries
 
     public string ProtocolItemID { get; set; } = null!;
 
-    public short SpecifiedUnitType { get; set; }
+    public int SpecifiedUnitType { get; set; }
 
     public byte IsDisplayAsVolume { get; set; }
